Moved EsyPackageJson out of files package. Dropped BaseFile parent class.
